refactor: Separate GitHub operations from plain Git operations
