+++ internal/command/command.go
@@ -29,6 +29,7 @@ import (
 	"time"
 
 	"github.com/googleapis/librarian/internal/container"
+	"github.com/googleapis/librarian/internal/githubrepo"
 	"github.com/googleapis/librarian/internal/gitrepo"
 	"github.com/googleapis/librarian/internal/statepb"
 	"github.com/googleapis/librarian/internal/utils"
@@ -311,20 +312,25 @@ func commitAll(repo *gitrepo.Repo, msg string) error {
 }
 
 // Push the contents of the language repo and create a new pull request.
-func pushAndCreatePullRequest(ctx *CommandContext, title string, description string) (*gitrepo.PullRequestMetadata, error) {
+func pushAndCreatePullRequest(ctx *CommandContext, title string, description string) (*githubrepo.PullRequestMetadata, error) {
 	if !flagPush {
 		return nil, nil
+	}
+
+	gitHubRepo, err := gitrepo.GetGitHubRepoFromRemote(ctx.languageRepo)
+	if err != nil {
+		return nil, err
 	}
 
 	// This should already have been validated to be non-empty by validatePush
 	gitHubAccessToken := os.Getenv(gitHubTokenEnvironmentVariable)
 	branch := fmt.Sprintf("librarian-%s", formatTimestamp(ctx.startTime))
-	err := gitrepo.PushBranch(ctx.languageRepo, branch, gitHubAccessToken)
+	err = gitrepo.PushBranch(ctx.languageRepo, branch, gitHubAccessToken)
 	if err != nil {
 		slog.Info(fmt.Sprintf("Received error pushing branch: '%s'", err))
 		return nil, err
 	}
-	pr, err := gitrepo.CreatePullRequest(ctx.ctx, ctx.languageRepo, branch, gitHubAccessToken, title, description)
+	pr, err := githubrepo.CreatePullRequest(ctx.ctx, gitHubRepo, branch, gitHubAccessToken, title, description)
 	if pr != nil {
 		return pr, err
 	}
